refactor: split GlobalStyles into styles.ts and utils.ts, migrate CTX to ColorPicker
