+++ src/components/PreviewText.tsx
@@ -1,9 +1,9 @@
 import React, { useState, useContext } from 'react';
 import { Text, StyleSheet } from 'react-native';
 import { runOnJS, useDerivedValue } from 'react-native-reanimated';
-import { CTX } from '../GlobalStyles';
 
 import type { PreviewTextProps } from '../types';
+import { CTX } from '../ColorPicker';
 
 export function PreviewText({ style = {}, colorFormat = 'hex' }: PreviewTextProps) {
   const { returnedResults, hueValue, saturationValue, brightnessValue, alphaValue } = useContext(CTX);
