+++ src/styles.ts
@@ -1,0 +1,49 @@
+import { I18nManager, StyleSheet } from 'react-native';
+
+export const styles = StyleSheet.create({
+  panel_container: {
+    position: 'relative',
+    alignSelf: 'stretch',
+    borderRadius: 5,
+  },
+  panel_image: {
+    position: 'absolute',
+    top: 0,
+    left: 0,
+    width: '100%',
+    height: '100%',
+    overflow: 'hidden',
+  },
+  handle: {
+    position: 'absolute',
+    ...(I18nManager.isRTL ? { right: 0 } : { left: 0 }),
+    top: 0,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  handleInner: {
+    width: '75%',
+    height: '75%',
+
+    shadowColor: '#000',
+    shadowOffset: {
+      width: 0,
+      height: 2,
+    },
+    shadowOpacity: 0.25,
+    shadowRadius: 3.84,
+
+    elevation: 5,
+  },
+  shadow: {
+    shadowColor: '#000',
+    shadowOffset: {
+      width: 0,
+      height: 2,
+    },
+    shadowOpacity: 0.25,
+    shadowRadius: 3.84,
+
+    elevation: 5,
+  },
+});
